# contracts/openapi.v0.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Svac.PublicApi | v0",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/me/sessions/{sessionId}": {
      "delete": {
        "tags": [
          "MeEndpoints"
        ],
        "operationId": "DeleteMeSession",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/v1/me/devices/{deviceId}": {
      "delete": {
        "tags": [
          "MeEndpoints"
        ],
        "operationId": "DeleteMeDevice",
        "parameters": [
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/v1/me/export": {
      "post": {
        "tags": [
          "ExportEndpoints"
        ],
        "operationId": "PostMeExport",
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportRequested"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitReached"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/export/{exportId}": {
      "get": {
        "tags": [
          "ExportEndpoints"
        ],
        "operationId": "GetMeExport",
        "parameters": [
          {
            "name": "exportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportStatus"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/v1/me/export/{exportId}/download": {
      "get": {
        "tags": [
          "ExportEndpoints"
        ],
        "operationId": "GetMeExportDownload",
        "parameters": [
          {
            "name": "exportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExportRequested": {
        "required": [
          "exportId"
        ],
        "type": "object",
        "properties": {
          "exportId": {
            "type": "string"
          }
        }
      },
      "ExportStatus": {
        "required": [
          "state",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "state": {
            "type": "string"
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "LimitReached": {
        "required": [
          "quotaKey",
          "messageKey",
          "resetsAt",
          "premiumExtends"
        ],
        "type": "object",
        "properties": {
          "quotaKey": {
            "type": "string"
          },
          "messageKey": {
            "type": "string"
          },
          "resetsAt": {
            "type": "string",
            "format": "date-time"
          },
          "premiumExtends": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Opaque, server-side, revocable session access token (SLICE_S3_CONTRACT.md §1b). Prefix \"sst_\", minted by /v1/signup/complete, /v1/auth/session, and /v1/auth/refresh. Header only: `Authorization: Bearer sst_...` — never a cookie or query parameter at S3. Carries zero decodable claims (never a JWT); revoked/expired/unknown tokens resolve to the anonymous actor server-side, never a distinct 401 oracle.",
        "scheme": "bearer",
        "bearerFormat": "sst_<256-bit-random>"
      }
    }
  },
...
}